rename test package
