+++ testutils/thriftarg2test/arg2_kv_builder_test.go
@@ -1,4 +1,4 @@
-package testutils
+package thriftarg2test
 
 import (
 	"testing"
